requirements.txt
newspaper3k
pandas
transformers
torch
Flask